Version provider packages Move current packages to <package-name>/v1 (backwards compatible with the non versioned provider)
