+++ auth/v1/credentials_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) OpenFaaS Author(s). All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-package auth
+package v1
 
 import (
 	"io/ioutil"
