+++ types/v1/requests.go
@@ -1,7 +1,7 @@
 // Copyright (c) Alex Ellis 2017. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-package types
+package v1
 
 type ScaleServiceRequest struct {
 	ServiceName string `json:"serviceName"`
